+++ src/cancelable.js
@@ -24,6 +24,7 @@ export {default as fetchResponse} from './cancelable/fetchResponse.js';
 export {default as fetchResponseIsoModule} from './cancelable/fetchResponseIsoModule.js';
 export {default as cancelify} from './cancelable/cancelify.js';
 export {default as cancelifyWithArityAbortable} from './cancelable/cancelifyAbortable.js';
+export {default as fromNodeCallbackWithArity} from './cancelable/fromNodeCallback.js';
 export {default as bi_tap} from './cancelable/biTap.js';
 export {default as biChain} from './cancelable/biChain.js';
 export {default as biMap} from './cancelable/biMap.js';