+++ src/either.d.ts
@@ -23,6 +23,8 @@ export function map<T, U>(fn: (x: T) => U, mx: Either<T>): Either<U>;
 /** @deprecated Use FL compliant utility function e.g. R.map */
 export function map<T, U>(fn: (x: T) => U) : (mx: Either<T>) => Either<U>;
 
+export function bimap<T, U>(fnLeft: (e: any) => any, fn: (x: T) => U): (mx: Either<T>) => Either<U>;
+
 /** @deprecated Use FL compliant utility function e.g. R.chain */
 export function chain<T, U>(factory: (x: T) => Either<U>, p: Either<T>) : Either<U>;
 /** @deprecated Use FL compliant utility function e.g. R.chain */